app/requirements.txt
asttokens==2.4.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
filelock==3.9.0
fsspec==2023.4.0
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
nvidia-cuda-runtime-cu11==2022.4.25
nvidia-cuda-runtime-cu117==11.7.60
nvidia-pyindex==1.0.9
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pywin32==306
pyzmq==25.1.1
requests==2.28.1
stack-data==0.6.3
sympy==1.12
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.8
